fixes in release script 18
