+++ gulpfile.js
@@ -10,7 +10,6 @@ gulp.task('default', (cb) => {
       yarn cmd package                      Packages Application in binaries
       yarn cmd config                       Upsert new NLU server config file
       yarn cmd bump                         Bump version and update changelog
-      yarn cmd release                      Create new release on both Git and Github 
     `)
   cb()
 })
@@ -18,4 +17,3 @@ gulp.task('default', (cb) => {
 gulp.task('package', package.package)
 gulp.task('config', config.upsertConfigFile)
 gulp.task('bump', release.bumpVersion)
-gulp.task('release', release.createNewRelease)
