+++ package.json
@@ -32,8 +32,6 @@
   },
   "devDependencies": {
     "conventional-changelog": "^3.1.24",
-    "gh-release": "^6.0.0",
-    "git-describe": "^4.0.4",
     "gulp": "^4.0.0",
     "gulp-cli": "^2.0.1",
     "pkg": "^4.5.1",
